Читал про паттерны. Изучал видео про чистую архитектуру. Много гугла, хабра и тд. Гуглил рум с перекрестными запросами, победил. Переношу записи для бд в эксель для удобства

+++ app/src/main/java/ru/sergeyzabelin/mylearning/ui/dictionary/DictionaryFragment.kt
@@ -46,7 +46,7 @@ class DictionaryFragment : Fragment() {
         findNavController().navigate(R.id.action_navDictionaryFragment_to_navDictionaryAddFragment)
     }
 
-    private fun onClickRecyclerItem(id: Int) {
+    private fun onClickRecyclerItem(id: Long) {
         findNavController().navigate(
             DictionaryFragmentDirections.actionNavDictionaryFragmentToNavDictionaryDetailFragment(
                 id
